requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
camelot-py==1.0.9
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
distro==1.9.0
et-xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.1
fsspec==2026.2.0
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.6.0
idna==3.11
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.11
langchain-core==1.2.18
langchain-text-splitters==1.1.1
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
openai==2.26.0
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
pdf2image==1.17.0
pdfminer-six==20260107
pillow==12.1.1
playwright==1.58.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pyee==13.0.1
pygments==2.19.2
pymupdf==1.27.2
pypdf==5.9.0
pypdfium2==5.6.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
zstandard==0.25.0
